refactor: Use logging/errors import path and update stripe api version
