+++ apps/web/src/integrations/orpc/router/categories.ts
@@ -15,7 +15,7 @@ import {
    updateCategorySchema,
 } from "@core/database/schemas/categories";
 import { user as userTable } from "@core/database/schemas/auth";
-import { WebAppError } from "@core/logging";
+import { WebAppError } from "@core/logging/errors";
 import { eq } from "drizzle-orm";
 import { z } from "zod";
 import { enqueueDeriveKeywordsWorkflow } from "@packages/workflows/workflows/derive-keywords-workflow";
