+++ apps/web/src/integrations/orpc/sdk/router/transactions.ts
@@ -14,7 +14,7 @@ import {
    ListTransactionsFilterSchema,
 } from "@montte/cli/contract";
 import dayjs from "dayjs";
-import { WebAppError } from "@core/logging";
+import { WebAppError } from "@core/logging/errors";
 import { emitFinanceTransactionCreated } from "@packages/events/finance";
 import { createBillableProcedure } from "../billable";
 import { sdkProcedure } from "../server";
