+++ apps/web/src/routes/api/$.ts
@@ -5,8 +5,8 @@ import { onError } from "@orpc/server";
 import { CompressionPlugin, RPCHandler } from "@orpc/server/fetch";
 import { BatchHandlerPlugin } from "@orpc/server/plugins";
 import { ZodToJsonSchemaConverter } from "@orpc/zod/zod4";
-import { FetchLoggingPlugin } from "@core/logging";
-import { getLogger } from "@core/logging";
+import { FetchLoggingPlugin } from "@core/logging/orpc-plugin";
+import { getLogger } from "@core/logging/root";
 import { createFileRoute } from "@tanstack/react-router";
 
 import router from "@/integrations/orpc/router";
