+++ apps/web/src/routes/api/rpc/$.ts
@@ -2,7 +2,7 @@ import "@/polyfill";
 
 import { RPCHandler } from "@orpc/server/fetch";
 import { BatchHandlerPlugin } from "@orpc/server/plugins";
-import { FetchLoggingPlugin } from "@core/logging";
+import { FetchLoggingPlugin } from "@core/logging/orpc-plugin";
 import { createFileRoute } from "@tanstack/react-router";
 import pino from "pino";
 import router from "@/integrations/orpc/router";
