+++ apps/web/src/server.ts
@@ -3,7 +3,7 @@ import "@/integrations/otel/init";
 
 import handler, { createServerEntry } from "@tanstack/react-start/server-entry";
 import { env } from "@core/environment/web";
-import { initLogger } from "@core/logging";
+import { initLogger } from "@core/logging/root";
 
 initLogger({ name: "montte", level: env.LOG_LEVEL });
