+++ package.json
@@ -207,7 +207,7 @@
       "db:push:prod": "bun run scripts/ensure-schemas.ts --env production && nx run core/database:db-push-prod",
       "db:studio:local": "nx run core/database:db-studio-local",
       "db:studio:prod": "nx run core/database:db-studio-prod",
-      "dev": "bun run scripts/dev-init.ts && nx run-many -t dev --parallel=3 -p web worker",
+      "dev": "bun run scripts/dev-init.ts; nx run-many -t dev --parallel=3 -p web worker",
       "dev:staging": "bun run scripts/check-env.ts --env staging && nx run-many -t dev:staging --parallel=3 -p web",
       "setup": "bun run scripts/setup.ts",
       "dev:all": "bun run scripts/dev-init.ts && nx run-many -t dev --parallel=6",
